Route to homepage on refresh
The issue here is that once the user has clicked and routed to the ProductPage, if the user refreshes the page it will lose access to the store, the product name and consequently the product itself.

Stop-gap solution is to send the user back to the home page in this scenario. The issues with this are discussed in journey.md